requirements.txt
certifi
chardet
idna
numpy
opencv-python
PyExecJS
requests
six
urllib3
python-dateutil
redis
pymongo
beautifulsoup4